# Greenovation-backend-main/docs/openapiGreenovation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Greenovation
  description: "This is the Open API documentation for the Greenovation API"
servers:
  - url: 'https://quason.tech/'
  - url: 'http://nattech.fib.upc.edu:40370/'
  - url: 'http://localhost:3000/'
paths:
...
  /edificis:
    post:
      tags:
        - Edificis
      security:
        - bearerAuth: []
      summary: Creates a building for the user bearer of the token
      operationId: postEdifici
      parameters:
        - name: default
          in: query
          description: 'If value equals true building will be built from default template'
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
            example: 'false'
        - name: num_cas
          in: query
          description: 'If default is not true the building will be created using the real life building identified with this parameter'
          schema:
            type: string
            example: 'WWR5FS9Z3'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  example: 'BUILDING FROM BODY'
                poblacio:
                  type: string
                  example: 'Barcelona'
                direccio:
                  type: string
                  example: 'Avinguda DIAGONAL 1'
                energia:
                  type: number
                  example: 100
                emissions:
                  type: number
                  example: 100
                ve:
                  type: boolean
                  example: true
                solar_fv:
                  type: boolean
                  example: true
                solar_ter:
                  type: boolean
                  example: true
                classificacio:
                  type: string
                  example: 'A'
      responses:
        201:
          description: Building created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edifici'
        400:
          description: 'Invalid parameters'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        401:
          description: 'Unauthorized to edit this building'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'No building found with this identifier'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    get:
      tags:
        - Edificis
      summary: Retrieves all Buildings
      operationId: getAllBuildings
      parameters:
        - name: lletra
          in: query
          description: 'Filter building by indicated leter'
          schema:
            type: string
            example: 'A'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Edifici'
  /edificis/{id}/aplicarMillora:
    post:
      tags:
        - Edificis
      summary: Applies an Upgrade to the building with the indicated id
      operationId: ApplyUpgrade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milloraId
          in: query
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/examples/message'
        400:
          description: 'Invalid parameters'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        404:
          description: 'No existeix edifici'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /edificis/{id}/aplicacions:
    get:
      tags:
        - Edificis
      summary: Gets all the applied Upgrades on the building with the indicated id
      operationId: getUpgradesOfABuilding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aplicacio'
        400:
          description: 'Invalid parameters in body'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        404:
          description: 'Aquest edifici no té millores aplicades'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /edificis/{id}/milloresAplicables:
    get:
      tags:
        - Edificis
      summary: Gets all the Upgrades that can be applied to the building with the indicated id
      operationId: getPossibleUpgradesOfABuilding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: categoriaId
          in: query
          required: true
          schema:
            type: integer
        - name: esDEdifici
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Millora'
        400:
          description: 'Invalid parameters in body'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        404:
          description: 'Aquest edifici no té millores aplicades'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /edificis/user/{userId}:
    get:
      tags:
        - Edificis
      summary: Gets the building that belongs to the user with the indicated id
      operationId: getBuildingByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edifici'
        404:
          description: 'No building found for this user'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    put:
      tags:
        - Edificis
      security:
        - bearerAuth: []
      summary: Edits the building that belongs to the user with the indicated id
      operationId: editBuildingByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  example: 'BUILDING FROM BODY'
                poblacio:
                  type: string
                  example: 'Barcelona'
                direccio:
                  type: string
                  example: 'Avinguda DIAGONAL 1'
                energia:
                  type: number
                  example: 100
                emissions:
                  type: number
                  example: 100
                ve:
                  type: boolean
                  example: true
                solar_fv:
                  type: boolean
                  example: true
                solar_ter:
                  type: boolean
                  example: true
                classificacio:
                  type: string
                  example: 'A'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edifici'
        400:
          description: 'Invalid parameters'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        401:
          description: 'Unauthorized to edit this building'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'Building with this userId not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    delete:
      tags:
        - Edificis
      security:
        - bearerAuth: []
      summary: Deletes the building that belongs to the user with the indicated id
      operationId: deleteBuildingByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'No content'
        401:
          description: 'Unauthorized to delete this building'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Invalid token'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'Building with this userId not found'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /edificis/recompensesRanking/{lletra}:
    get:
      tags:
        - Edificis
      summary: Gets the ranking rewards of the indicated certificate
      operationId: getRankingRewardsByCertificate
      parameters:
        - name: lletra
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recompensa'
        404:
          description: 'No building found for this user'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
...
components:
  schemas:
...
    Edifici:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nom:
          type: string
          example: 'WWR5FS9Z3'
        poblacio:
          type: string
          example: 'Sant Gregori'
        direccio:
          type: string
          example: 'Carrer FRANCESC DE VERNTALLAT 20'
        energia:
          type: number
          example: 35.67
        emissions:
          type: number
          example: 6.95
        ve:
          type: boolean
          example: true
        solar_fv:
          type: boolean
          example: true
        solar_ter:
          type: boolean
          example: true
        classificacio:
          type: string
          example: 'A'
        userId:
          type: integer
          example: 1
        felicitatVeins:
          type: integer
          example: 75
        dinersAlMes:
          type: integer
          example: 100
        energiaIni:
          type: number
          example: 35.67
        classificacioIni:
          type: string
          example: 'A'
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
...
    Millora:
      type: object
      properties:
        id:
          type: number
          example: 1
        nomCat:
          type: string
          example: 'Millora de la il·luminació amb LEDs.'
        nomEsp:
          type: string
          example: 'Mejora de la iluminación con LEDs.'
        nomEng:
          type: string
          example: 'LED Lighting Upgrade'
        descripcioCat:
          type: string
          example: 'Substituir totes les llums incandescents i fluorescents per llums LEDs.'
        descripcioEsp:
          type: string
          example: 'Sustituir todas las luces incandescentes y fluorescentes por luces LEDs.'
        descripcioEng:
          type: string
          example: 'Replace all old incandescent and fluorescent lights with energy-efficient LED lights.'
        reasoningCat:
          type: string
          example: 'Els LED consumeixen menys electricitat i duren més, el que redueix els costs de manteniment i ús general energètic del teu edifici.'
        reasoningEsp:
          type: string
          example: 'Los LED consumen menos electricidad y duran más, lo que reduce el uso general de energía y los costes de mantenimiento del edificio.'
        reasoningEng:
          type: string
          example: 'LEDs consume less electricity and last longer, reducing the building’s overall energy usage and maintenance costs.'
        cost:
          type: number
          example: 2000
        reduccioEnergia:
          type: number
          example: 5
        reduccioCO2:
          type: number
          example: 0
        milloraFelicitat:
          type: number
          example: 0
        esDEdifici:
          type: boolean
          example: true
        tempsDObra:
          type: number
          example: 0
        categoriaId:
          type: number
          example: 1
        lletraId:
          type: number
          example: 1
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
    Aplicacio:
      type: object
      properties:
        edificiId:
          type: number
          example: 1
        milloraId:
          type: number
          example: 1
        totalCompres:
          type: number
          example: 1
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
    Recompensa:
      type: object
      properties:
        id:
          type: number
          example: 1
        posicio:
          type: string
          example: "1st"
        recompensa:
          type: number
          example: 100
        createdAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
        updatedAt:
          type: string
          example: '2023-11-22T19:29:35.963Z'
...
  examples:
    message:
      value: 'Successful'
    error400:
      value:
        message: 'Invalid information entered'
    error401:
      value:
        message: 'Unauthorized to perform this operation'
    error403:
      value:
        message: 'Invalid token'
    error404:
      value:
        message: 'Not found'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT